Refactor alloc module and add TryClone
