+++ core/alloc/collections.rs
@@ -1,241 +1,9 @@
-//! Turso-owned allocation namespace.
-//!
-//! Stable builds use `allocator-api2` where it has allocator-aware types and
-//! fall back to `std` for collections that do not have stable allocator-aware
-//! equivalents. Builds compiled with `--cfg nightly` use Rust's unstable
-//! `allocator_api` collection parameters.
+use std::hash::{BuildHasher, Hash};
 
-use std::{
-    fmt,
-    hash::{BuildHasher, Hash},
-    ptr::NonNull,
-    sync::OnceLock,
+use super::{
+    AllocError, Arc, BTreeMap, BTreeSet, BinaryHeap, Box, HashMap, HashSet, LinkedList, Rc,
+    TryReserveError, TursoAllocator, Vec, VecDeque,
 };
-
-#[cfg(not(nightly))]
-mod api {
-    pub use allocator_api2::{
-        alloc::{AllocError, Allocator as ApiAllocator, Global, Layout},
-        collections::TryReserveError,
-    };
-}
-
-#[cfg(nightly)]
-mod api {
-    pub use std::alloc::{AllocError, Allocator as ApiAllocator, Global, Layout};
-}
-
-pub use api::{AllocError, Layout};
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct TryReserveError;
-
-impl fmt::Display for TryReserveError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.write_str("memory allocation failed")
-    }
-}
-
-impl std::error::Error for TryReserveError {}
-
-#[cfg(not(nightly))]
-impl From<api::TryReserveError> for TryReserveError {
-    fn from(_: api::TryReserveError) -> Self {
-        Self
-    }
-}
-
-impl From<std::collections::TryReserveError> for TryReserveError {
-    fn from(_: std::collections::TryReserveError) -> Self {
-        Self
-    }
-}
-
-#[derive(Clone, Copy, Debug, Default)]
-pub struct TursoAllocator;
-
-pub type Allocator = TursoAllocator;
-
-#[cfg(not(nightly))]
-pub type Box<T> = allocator_api2::boxed::Box<T, TursoAllocator>;
-#[cfg(nightly)]
-pub type Box<T> = std::boxed::Box<T, TursoAllocator>;
-
-#[cfg(not(nightly))]
-pub type Vec<T> = allocator_api2::vec::Vec<T, TursoAllocator>;
-#[cfg(nightly)]
-pub type Vec<T> = std::vec::Vec<T, TursoAllocator>;
-
-pub use crate::{__turso_alloc_try_vec as try_vec, __turso_alloc_vec as vec};
-
-#[doc(hidden)]
-#[macro_export]
-macro_rules! __turso_alloc_vec_count {
-    ($($element:expr),*) => {
-        <[()]>::len(&[$($crate::__turso_alloc_vec_count!(@sub $element)),*])
-    };
-    (@sub $element:expr) => {
-        ()
-    };
-}
-
-#[doc(hidden)]
-#[macro_export]
-macro_rules! __turso_alloc_vec {
-    () => {
-        <$crate::alloc::Vec<_> as $crate::alloc::TursoAllocExt>::new()
-    };
-    ($element:expr; $count:expr) => {{
-        let count = $count;
-        let mut values =
-            <$crate::alloc::Vec<_> as $crate::alloc::TursoVecExt<_>>::with_capacity(count);
-        values.resize(count, $element);
-        values
-    }};
-    ($($element:expr),+ $(,)?) => {{
-        let mut values =
-            <$crate::alloc::Vec<_> as $crate::alloc::TursoVecExt<_>>::with_capacity(
-                $crate::__turso_alloc_vec_count!($($element),+),
-            );
-        $(values.push($element);)+
-        values
-    }};
-}
-
-#[doc(hidden)]
-#[macro_export]
-macro_rules! __turso_alloc_try_vec {
-    () => {
-        Ok::<_, $crate::alloc::TryReserveError>(
-            <$crate::alloc::Vec<_> as $crate::alloc::TursoAllocExt>::new(),
-        )
-    };
-    ($element:expr; $count:expr) => {{
-        (|| {
-            let count = $count;
-            let mut values =
-                <$crate::alloc::Vec<_> as $crate::alloc::TursoTryWithCapacityExt>::try_with_capacity(
-                    count,
-                )?;
-            values.resize(count, $element);
-            Ok::<_, $crate::alloc::TryReserveError>(values)
-        })()
-    }};
-    ($($element:expr),+ $(,)?) => {{
-        (|| {
-            let mut values =
-                <$crate::alloc::Vec<_> as $crate::alloc::TursoTryWithCapacityExt>::try_with_capacity(
-                    $crate::__turso_alloc_vec_count!($($element),+),
-                )?;
-            $(values.try_push($element)?;)+
-            Ok::<_, $crate::alloc::TryReserveError>(values)
-        })()
-    }};
-}
-
-pub type String = std::string::String;
-
-pub type HashMap<K, V, S = rustc_hash::FxBuildHasher> = std::collections::HashMap<K, V, S>;
-
-pub type HashSet<T, S = rustc_hash::FxBuildHasher> = std::collections::HashSet<T, S>;
-
-#[cfg(not(nightly))]
-pub type BTreeMap<K, V> = std::collections::BTreeMap<K, V>;
-#[cfg(nightly)]
-pub type BTreeMap<K, V> = std::collections::BTreeMap<K, V, TursoAllocator>;
-
-#[cfg(not(nightly))]
-pub type BTreeSet<T> = std::collections::BTreeSet<T>;
-#[cfg(nightly)]
-pub type BTreeSet<T> = std::collections::BTreeSet<T, TursoAllocator>;
-
-#[cfg(not(nightly))]
-pub type VecDeque<T> = std::collections::VecDeque<T>;
-#[cfg(nightly)]
-pub type VecDeque<T> = std::collections::VecDeque<T, TursoAllocator>;
-
-#[cfg(not(nightly))]
-pub type BinaryHeap<T> = std::collections::BinaryHeap<T>;
-#[cfg(nightly)]
-pub type BinaryHeap<T> = std::collections::BinaryHeap<T, TursoAllocator>;
-
-#[cfg(not(nightly))]
-pub type LinkedList<T> = std::collections::LinkedList<T>;
-#[cfg(nightly)]
-pub type LinkedList<T> = std::collections::LinkedList<T, TursoAllocator>;
-
-// TODO: design allocator-aware shared-pointer support that still preserves
-// shuttle's deterministic sync behavior.
-pub type Arc<T> = crate::sync::Arc<T>;
-pub type Weak<T> = crate::sync::Weak<T>;
-
-#[cfg(not(nightly))]
-pub type Rc<T> = std::rc::Rc<T>;
-#[cfg(nightly)]
-pub type Rc<T> = std::rc::Rc<T, TursoAllocator>;
-
-#[cfg(not(nightly))]
-pub type RcWeak<T> = std::rc::Weak<T>;
-#[cfg(nightly)]
-pub type RcWeak<T> = std::rc::Weak<T, TursoAllocator>;
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum SetAllocatorError {
-    AlreadyInitialized,
-}
-
-impl fmt::Display for SetAllocatorError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::AlreadyInitialized => f.write_str("Turso allocator is already initialized"),
-        }
-    }
-}
-
-impl std::error::Error for SetAllocatorError {}
-
-/// Backend for Turso heap allocations.
-///
-/// # Safety
-///
-/// Implementations must uphold the `Allocator` contract for every allocation
-/// returned from `allocate`, including zero-sized layouts.
-pub unsafe trait TursoAllocBackend: Sync {
-    fn allocate(&self, layout: Layout) -> Result<NonNull<[u8]>, AllocError>;
-
-    /// # Safety
-    ///
-    /// `ptr` and `layout` must describe a live block previously returned by
-    /// this backend.
-    unsafe fn deallocate(&self, ptr: NonNull<u8>, layout: Layout);
-}
-
-struct DefaultBackend;
-
-unsafe impl TursoAllocBackend for DefaultBackend {
-    fn allocate(&self, layout: Layout) -> Result<NonNull<[u8]>, AllocError> {
-        <api::Global as api::ApiAllocator>::allocate(&api::Global, layout)
-    }
-
-    unsafe fn deallocate(&self, ptr: NonNull<u8>, layout: Layout) {
-        unsafe {
-            <api::Global as api::ApiAllocator>::deallocate(&api::Global, ptr, layout);
-        }
-    }
-}
-
-static DEFAULT_BACKEND: DefaultBackend = DefaultBackend;
-static BACKEND: OnceLock<&'static dyn TursoAllocBackend> = OnceLock::new();
-
-pub fn set_allocator(backend: &'static dyn TursoAllocBackend) -> Result<(), SetAllocatorError> {
-    BACKEND
-        .set(backend)
-        .map_err(|_| SetAllocatorError::AlreadyInitialized)
-}
-
-fn backend() -> &'static dyn TursoAllocBackend {
-    BACKEND.get().copied().unwrap_or(&DEFAULT_BACKEND)
-}
 
 fn vec<T>() -> Vec<T> {
     Vec::new_in(TursoAllocator)
@@ -856,247 +624,5 @@ impl<T> TursoNewExt<T> for Rc<T> {
 impl<T> TursoNewExt<T> for Rc<T> {
     fn new(value: T) -> Self {
         rc(value)
-    }
-}
-
-#[cfg(not(nightly))]
-unsafe impl api::ApiAllocator for TursoAllocator {
-    fn allocate(&self, layout: Layout) -> Result<NonNull<[u8]>, AllocError> {
-        backend().allocate(layout)
-    }
-
-    unsafe fn deallocate(&self, ptr: NonNull<u8>, layout: Layout) {
-        unsafe {
-            backend().deallocate(ptr, layout);
-        }
-    }
-}
-
-#[cfg(nightly)]
-unsafe impl api::ApiAllocator for TursoAllocator {
-    fn allocate(&self, layout: Layout) -> Result<NonNull<[u8]>, AllocError> {
-        backend().allocate(layout)
-    }
-
-    unsafe fn deallocate(&self, ptr: NonNull<u8>, layout: Layout) {
-        unsafe {
-            backend().deallocate(ptr, layout);
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    struct LowerBoundOnly {
-        next: usize,
-        end: usize,
-    }
-
-    impl Iterator for LowerBoundOnly {
-        type Item = usize;
-
-        fn next(&mut self) -> Option<Self::Item> {
-            if self.next == self.end {
-                return None;
-            }
-            let value = self.next;
-            self.next += 1;
-            Some(value)
-        }
-
-        fn size_hint(&self) -> (usize, Option<usize>) {
-            (0, None)
-        }
-    }
-
-    #[test]
-    fn try_extend_accepts_exact_size_iterators() {
-        let mut values = Vec::new();
-
-        values.try_extend([1, 2, 3]).unwrap();
-
-        assert_eq!(values.as_slice(), &[1, 2, 3]);
-    }
-
-    #[test]
-    fn try_extend_accepts_iterators_without_upper_bounds() {
-        let mut values = Vec::new();
-
-        values
-            .try_extend(LowerBoundOnly { next: 0, end: 3 })
-            .unwrap();
-
-        assert_eq!(values.as_slice(), &[0, 1, 2]);
-    }
-
-    #[test]
-    fn hash_map_try_insert_and_extend_reserve_before_mutation() {
-        let mut values: HashMap<&str, usize> = TursoAllocExt::new();
-
-        assert_eq!(
-            TursoHashMapExt::try_insert(&mut values, "one", 1).unwrap(),
-            None
-        );
-        assert_eq!(
-            TursoHashMapExt::try_insert(&mut values, "one", 11).unwrap(),
-            Some(1)
-        );
-        values.try_extend([("two", 2), ("three", 3)]).unwrap();
-
-        assert_eq!(values.get("one"), Some(&11));
-        assert_eq!(values.get("two"), Some(&2));
-        assert_eq!(values.get("three"), Some(&3));
-    }
-
-    #[test]
-    fn hash_set_try_insert_and_extend_reserve_before_mutation() {
-        let mut values: HashSet<usize> = TursoAllocExt::new();
-
-        assert!(TursoHashSetExt::try_insert(&mut values, 1).unwrap());
-        assert!(!TursoHashSetExt::try_insert(&mut values, 1).unwrap());
-        values.try_extend([2, 3]).unwrap();
-
-        assert!(values.contains(&1));
-        assert!(values.contains(&2));
-        assert!(values.contains(&3));
-    }
-
-    #[test]
-    fn vec_deque_try_push_and_extend_reserve_before_mutation() {
-        let mut values: VecDeque<usize> = TursoAllocExt::new();
-
-        values.try_push_back(2).unwrap();
-        values.try_push_front(1).unwrap();
-        values.try_extend([3, 4]).unwrap();
-
-        assert_eq!(values.pop_front(), Some(1));
-        assert_eq!(values.pop_front(), Some(2));
-        assert_eq!(values.pop_front(), Some(3));
-        assert_eq!(values.pop_front(), Some(4));
-        assert_eq!(values.pop_front(), None);
-    }
-
-    #[test]
-    fn binary_heap_try_push_and_extend_reserve_before_mutation() {
-        let mut values: BinaryHeap<usize> = TursoAllocExt::new();
-
-        values.try_push(2).unwrap();
-        values.try_extend([1, 3]).unwrap();
-
-        assert_eq!(values.pop(), Some(3));
-        assert_eq!(values.pop(), Some(2));
-        assert_eq!(values.pop(), Some(1));
-        assert_eq!(values.pop(), None);
-    }
-
-    #[test]
-    fn iterator_try_collect_builds_turso_vec() {
-        let values: Vec<_> = [1, 2, 3].into_iter().try_collect().unwrap();
-
-        assert_eq!(values.as_slice(), &[1, 2, 3]);
-    }
-
-    #[test]
-    fn try_vec_macro_builds_turso_vecs() {
-        let empty: Vec<usize> = try_vec![].unwrap();
-        let repeated: Vec<_> = try_vec![7; 3].unwrap();
-        let listed: Vec<_> = try_vec![1, 2, 3].unwrap();
-
-        assert!(empty.is_empty());
-        assert_eq!(repeated.as_slice(), &[7, 7, 7]);
-        assert_eq!(listed.as_slice(), &[1, 2, 3]);
-    }
-
-    #[test]
-    fn iterator_try_collect_builds_turso_collections() {
-        let map: HashMap<_, _> = [("one", 1), ("two", 2)].into_iter().try_collect().unwrap();
-        let set: HashSet<_> = [1, 2, 3].into_iter().try_collect().unwrap();
-        let heap: BinaryHeap<_> = [1, 3, 2].into_iter().try_collect().unwrap();
-
-        assert_eq!(map.get("one"), Some(&1));
-        assert!(set.contains(&3));
-        assert_eq!(heap.peek(), Some(&3));
-    }
-
-    #[test]
-    fn iterator_try_collect_builds_option_collection() {
-        let values: Option<Vec<_>> = [Some(1), Some(2), Some(3)]
-            .into_iter()
-            .try_collect()
-            .unwrap();
-        let none: Option<Vec<_>> = [Some(1), None, Some(3)].into_iter().try_collect().unwrap();
-
-        assert_eq!(values.unwrap().as_slice(), &[1, 2, 3]);
-        assert!(none.is_none());
-    }
-
-    #[test]
-    fn iterator_try_collect_builds_result_collection() {
-        let values: Result<Vec<_>, &str> = [Ok(1), Ok(2), Ok(3)].into_iter().try_collect().unwrap();
-        let error: Result<Vec<_>, &str> = [Ok(1), Err("bad"), Ok(3)]
-            .into_iter()
-            .try_collect()
-            .unwrap();
-
-        assert_eq!(values.unwrap().as_slice(), &[1, 2, 3]);
-        assert_eq!(error, Err("bad"));
-    }
-
-    #[test]
-    fn iterator_try_collect_converts_result_error() {
-        #[derive(Debug, PartialEq)]
-        struct Converted(&'static str);
-
-        impl From<&'static str> for Converted {
-            fn from(value: &'static str) -> Self {
-                Self(value)
-            }
-        }
-
-        let values: Result<Vec<_>, Converted> = [
-            Ok::<_, &'static str>(1),
-            Ok::<_, &'static str>(2),
-            Ok::<_, &'static str>(3),
-        ]
-        .into_iter()
-        .try_collect::<Result<Vec<_>, Converted>>()
-        .unwrap();
-        let error: Result<Vec<_>, Converted> = [Ok(1), Err("bad"), Ok(3)]
-            .into_iter()
-            .try_collect::<Result<Vec<_>, Converted>>()
-            .unwrap();
-
-        assert_eq!(values.unwrap().as_slice(), &[1, 2, 3]);
-        assert_eq!(error, Err(Converted("bad")));
-    }
-
-    #[test]
-    fn iterator_try_collect_accepts_try_vec_results() {
-        let values: crate::Result<Vec<_>> = [1usize, 2]
-            .into_iter()
-            .map(|count| try_vec![false; count])
-            .try_collect::<crate::Result<Vec<_>>>()
-            .unwrap();
-        let values = values.unwrap();
-
-        assert_eq!(values[0].as_slice(), &[false]);
-        assert_eq!(values[1].as_slice(), &[false, false]);
-    }
-
-    #[test]
-    fn try_with_capacity_builds_turso_collections() {
-        let values: Vec<usize> = TursoTryWithCapacityExt::try_with_capacity(3).unwrap();
-        let map: HashMap<usize, usize> = TursoTryWithCapacityExt::try_with_capacity(3).unwrap();
-        let set: HashSet<usize> = TursoTryWithCapacityExt::try_with_capacity(3).unwrap();
-        let queue: VecDeque<usize> = TursoTryWithCapacityExt::try_with_capacity(3).unwrap();
-        let heap: BinaryHeap<usize> = TursoTryWithCapacityExt::try_with_capacity(3).unwrap();
-
-        assert!(values.capacity() >= 3);
-        assert!(map.capacity() >= 3);
-        assert!(set.capacity() >= 3);
-        assert!(queue.capacity() >= 3);
-        assert!(heap.capacity() >= 3);
     }
 }
